sdk/python/pyproject.toml
[project]
name = "unified-push"
version = "1.0.0"
description = "End-to-end encrypted unified push notification SDK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "eciespy>=0.4.2",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9.0",
]

[project.scripts]
unified-push = "unified_push.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
